requirements_unique.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile ./requirements_unique.in
#
aiofiles==24.1.0
    # via -r ./requirements_unique.in
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.11.2
    # via -r ./requirements_unique.in
aiosignal==1.3.1
    # via aiohttp
astropy==6.1.6
    # via -r ./requirements_unique.in
astropy-iers-data==0.2024.11.11.0.32.38
    # via astropy
async-timeout==5.0.1
    # via aiohttp
attrs==24.2.0
    # via aiohttp
beautifulsoup4==4.12.3
    # via
    #   -r ./requirements_unique.in
    #   gdown
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.16.1
    # via
    #   gdown
    #   torch
    #   triton
fonttools==4.55.0
    # via matplotlib
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.10.0
    # via torch
gdown==5.2.0
    # via -r ./requirements_unique.in
idna==3.10
    # via
    #   requests
    #   yarl
imageio==2.36.0
    # via scikit-image
jinja2==3.1.4
    # via torch
kiwisolver==1.4.7
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
markupsafe==3.0.2
    # via jinja2
matplotlib==3.9.2
    # via -r ./requirements_unique.in
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via
    #   scikit-image
    #   torch
numpy==2.1.3
    # via
    #   -r ./requirements_unique.in
    #   astropy
    #   contourpy
    #   imageio
    #   matplotlib
    #   pandas
    #   pyerfa
    #   scikit-image
    #   scipy
    #   tifffile
    #   torchvision
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
packaging==24.2
    # via
    #   astropy
    #   lazy-loader
    #   matplotlib
    #   scikit-image
pandas==2.2.3
    # via -r ./requirements_unique.in
pillow==11.0.0
    # via
    #   -r ./requirements_unique.in
    #   imageio
    #   matplotlib
    #   scikit-image
    #   torchvision
propcache==0.2.0
    # via
    #   aiohttp
    #   yarl
pyerfa==2.0.1.5
    # via astropy
pyparsing==3.2.0
    # via matplotlib
pysocks==1.7.1
    # via requests
python-dateutil==2.9.0.post0
    # via
    #   -r ./requirements_unique.in
    #   matplotlib
    #   pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via astropy
requests[socks]==2.32.3
    # via gdown
scikit-image==0.24.0
    # via -r ./requirements_unique.in
scipy==1.14.1
    # via
    #   -r ./requirements_unique.in
    #   scikit-image
six==1.16.0
    # via python-dateutil
soupsieve==2.6
    # via beautifulsoup4
sympy==1.13.1
    # via torch
tifffile==2024.9.20
    # via scikit-image
torch==2.5.1
    # via
    #   -r ./requirements_unique.in
    #   torchvision
torchvision==0.20.1
    # via -r ./requirements_unique.in
tqdm==4.67.0
    # via gdown
triton==3.1.0
    # via torch
typing-extensions==4.12.2
    # via
    #   multidict
    #   torch
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
yarl==1.17.1
    # via aiohttp
